% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDissim.R
\name{discretizeQuant}
\alias{discretizeQuant}
\title{Descretize Quantitative Variables.}
\usage{
discretizeQuant(myDataQuant, noice = TRUE)
}
\arguments{
\item{myDataQuant}{A data frame which includes quantitative variables in columns.}

\item{noice}{Noice indicator. If noice = TRUE data standerdization is done by
deviding the difference between data point and median of the variable by the range of the variable.
If noice = FALSE data standerdization is done by deviding the difference between data point and
mean of the variable by the standard deviation of the variable.}
}
\value{
A data frame consists of discretized quantitative variables.
}
\description{
Takes in a data frame which contains only Quantitative varables in columns. Standadize the variables.
Discretize quantitative variables and returns discretized quantitative variables. Discretization was performed by
equal width bining algorithm.
}
\examples{
QuantVars <- data.frame(Qnvar1 = c(1.5,3.2,4.9,5), Qnvar2 = c(4.8,2,1.1,5.8))
Discretized <- discretizeQuant(QuantVars)
}

